"Refactor simulation structure and modularize WalkType logic
Consolidated WalkType and WalkParams into a new simulation model module for enhanced modularity and maintainability. Removed redundant code, reorganized utilities, and updated traits and steps to reflect the new structure. This cleanup simplifies simulation workflows and improves code clarity."
